complete dispatcher, pusher

+++ model/Quiz.go
@@ -5,8 +5,8 @@ import "go.mongodb.org/mongo-driver/bson/primitive"
 type Stat struct {
 	Difficult string `json:"difficult" bson:"difficult"`
 	Author    string `json:"author" bson:"author"`
-	Submitted int64  `json:"submitted" bson:"submitted"`
-	Score     int64  `json:"score" bson:"score"`
+	Submitted int32  `json:"submitted" bson:"submitted"`
+	Score     int32  `json:"score" bson:"score"`
 }
 
 type Template struct {
@@ -15,9 +15,9 @@ type Template struct {
 }
 
 type Limit struct {
-	Submit  int64 `json:"submit" bson:"submit"`
+	Submit  int32 `json:"submit" bson:"submit"`
 	Memory  int64 `json:"memory" bson:"memory"`
-	Cpu     int64 `json:"cpu" bson:"cpu"`
+	Cpu     int32 `json:"cpu" bson:"cpu"`
 	Timeout int64 `json:"timeout" bson:"timeout"`
 }
